Force the computation of the amounts fields on sale.order
By writing again on sale.order.line

The amounts (amount_total, ...) computed fields on 'sale.order' are not
triggered when magento.sale.order.line are created.  It might be a v8
regression, because they were triggered in v7. Before getting a better
correction, force the computation by writing again on the line.

It fixes 2 tests.